refactor: 홈 화면 관련 UI 모델 패키지 구조 개선
* **refactor: UI 모델 클래스 패키지 이동**
    * `LineupItemUiModel`, `LineUpItemOfDayUiModel`, `LineUpItemGroupUiModel` 등 홈 화면에서 사용하는 UI 모델 클래스들을 `presentation.home` 패키지에서 하위 패키지인 `presentation.home.model`로 이동했습니다.
    * 패키지 변경에 따라 `HomeViewModel`, `HomeScreen`, `HomeLineupItem` 및 관련 테스트 코드의 임포트 경로를 수정했습니다.

* **refactor: LineupUiState 의존성 수정**
    * `LineupUiState`에서 이동된 패키지의 `LineUpItemOfDayUiModel`을 참조하도록 수정했습니다.
