+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/home/LineupUiState.kt
@@ -1,5 +1,7 @@
 package com.daedan.festabook.presentation.home
 
+import com.daedan.festabook.presentation.home.model.LineUpItemOfDayUiModel
+
 sealed interface LineupUiState {
     data object Loading : LineupUiState
